Moved pipeline utils to common folder
