+++ packages/pipe-vms-ingestion/vms_ingestion/normalization/__init__.py
@@ -1,37 +1,14 @@
-from apache_beam.options.pipeline_options import PipelineOptions, StandardOptions
-from apache_beam.runners import PipelineState
+from common.pipeline import (
+    build_pipeline_options_with_defaults,
+    is_blocking_run,
+    success_states,
+)
+from logger import logger
 from vms_ingestion.normalization.options import NormalizationOptions
 from vms_ingestion.normalization.pipeline import NormalizationPipeline
 
-from logger import logger
-
 logger.setup_logger(1)
 logging = logger.get_logger()
-
-
-def build_pipeline_options_with_defaults(argv):
-    return PipelineOptions(
-        flags=argv,
-    )
-
-
-def is_blocking_run(pipeline_options):
-    return (
-        pipeline_options.view_as(NormalizationOptions).wait_for_job
-        or pipeline_options.view_as(StandardOptions).runner == "DirectRunner"
-    )
-
-
-def success_states(pipeline_options):
-    if is_blocking_run(pipeline_options):
-        return {PipelineState.DONE}
-    else:
-        return {
-            PipelineState.DONE,
-            PipelineState.RUNNING,
-            PipelineState.UNKNOWN,
-            PipelineState.PENDING,
-        }
 
 
 def run_normalization(argv):
@@ -44,13 +21,13 @@ def run_normalization(argv):
     pipeline = NormalizationPipeline(options=options)
     result = pipeline.run()
 
-    if is_blocking_run(options):
+    if is_blocking_run(options, NormalizationOptions):
         logging.info("Waiting for job to finish")
         result.wait_until_finish()
 
     logging.info("Pipeline launch result %s", result.state)
 
-    if result.state in success_states(options):
+    if result.state in success_states(options, NormalizationOptions):
         logging.info("Terminating process successfully")
         exit(0)
     else:
